complete implementation of storage
